+++ src/Header/Header.tsx
@@ -1,16 +1,22 @@
-import React, { forwardRef } from 'react'
+import React, { forwardRef, useState } from 'react'
 
 import { cn } from '@/__private__/utils/bem'
-import { HeaderProps, HeaderComponent } from './types'
+import { HeaderProps, HeaderComponent, DefaultItemLanguages } from './types'
 import { Layout } from '@/Layout'
 import { Menu } from '@/Menu'
 import { HeaderLogo } from './HeaderLogo'
 import { HeaderLogin } from './HeaderLogin'
+import { HeaderSearch } from './HeaderSearch'
 import { Notifications } from '@/Notifications'
 import { TileMenu } from '@/TileMenu'
 import { useBreakpoints } from '@consta/uikit/useBreakpoints'
 import { MobileMenu } from '@/MobileMenu'
 import { Breadcrumbs } from '@consta/uikit/BreadcrumbsCanary'
+import { SelectMenu } from '@/SelectMenu'
+import { ButtonMenu } from '@/ButtonMenu'
+import { Languages, LanguagesProps } from '@/Languages'
+
+import { getMobileMenu } from './helpers'
 
 import './Header.css'
 
@@ -75,7 +81,7 @@ const HeaderRender = (props: HeaderProps, ref: React.Ref<HTMLDivElement>) => {
     getTileMenuItemHref,
     getTileMenuItemImage,
     getTileMenuItemOnClick,
-    getTileMenuItemTitle,
+    getTileMenuItemLabel,
     tileMenuTitle,
 
     // Breadcrumbs
@@ -88,12 +94,64 @@ const HeaderRender = (props: HeaderProps, ref: React.Ref<HTMLDivElement>) => {
     getBreadcrumbsItemOnClick,
     breadcrumbsLastItemIsLink,
     breadcrumbsFitMode,
+
+    // Search
+    searchValue,
+    searchOnChange,
+    searchOnSubmit,
+    searchPlaceholder,
+
+    // SecondaryMenu
+    secondaryMenuLabel,
+    secondaryMenu,
+    getSecondaryMenuItemHref,
+    getSecondaryMenuItemLabel,
+    getSecondaryMenuItemOnClick,
+    getSecondaryMenuItemSubMenu,
+    getSecondaryMenuItemTarget,
+
+    // SocialMedia
+    socialMedia,
+    getSocialMediaItemHref,
+    getSocialMediaItemIcon,
+    getSocialMediaItemLabel,
+    getSocialMediaItemOnClick,
+    getSocialMediaItemTarget,
+    onSocialMediaItemClick,
+
+    // Languages
+    languages,
+    languageValue: languageValueProp,
+    getLanguagesItemHref,
+    getLanguagesItemLabel,
+    getLanguagesItemOnClick,
+    getLanguagesItemTarget,
+    onLanguageChange: onLanguageChangeProp,
+
+    // AdditionalButtons
+    additionalButtons,
+    getAdditionalButtonsItemHref,
+    getAdditionalButtonsItemIcon,
+    getAdditionalButtonsItemLabel,
+    getAdditionalButtonsItemOnClick,
+    getAdditionalButtonsItemTarget,
+    onAdditionalButtonsItemClick,
+
     ...otherProps
   } = props
 
   const breakpoints = useBreakpoints({ s: 800, m: 1200 })
 
+  const [languagesValue, setLanguagesValue] = useState<DefaultItemLanguages | undefined>(
+    languageValueProp || languages?.[0]
+  )
+
   const elementZIndex = typeof props.style?.zIndex === 'number' ? props.style.zIndex : undefined
+
+  const onLanguageChange: LanguagesProps<DefaultItemLanguages>['onChange'] = params => {
+    onLanguageChangeProp?.(params)
+    setLanguagesValue(params.item)
+  }
 
   if (breakpoints.m) {
     return (
@@ -101,10 +159,87 @@ const HeaderRender = (props: HeaderProps, ref: React.Ref<HTMLDivElement>) => {
         {...otherProps}
         ref={ref}
         className={cnHeader({ fixed }, [className])}
+        rowTop={{
+          left:
+            secondaryMenuLabel && secondaryMenu?.length ? (
+              <SelectMenu
+                label={secondaryMenuLabel}
+                items={secondaryMenu}
+                getItemHref={getSecondaryMenuItemHref}
+                getItemLabel={getSecondaryMenuItemLabel}
+                getItemOnClick={getSecondaryMenuItemOnClick}
+                getItemSubMenu={getSecondaryMenuItemSubMenu}
+                getItemTarget={getSecondaryMenuItemTarget}
+                style={{ zIndex: elementZIndex }}
+              />
+            ) : (
+              undefined
+            ),
+          center: socialMedia?.length ? (
+            <ButtonMenu
+              items={socialMedia}
+              getItemHref={getSocialMediaItemHref}
+              getItemIcon={getSocialMediaItemIcon}
+              getItemLabel={getSocialMediaItemLabel}
+              getItemOnClick={getSocialMediaItemOnClick}
+              getItemTarget={getSocialMediaItemTarget}
+              onItemClick={onSocialMediaItemClick}
+              view="clear"
+              onlyIcon
+              size="s"
+            />
+          ) : (
+            undefined
+          ),
+          right: (
+            <div className={cnHeader('RowTopRight')}>
+              {[
+                languages?.length ? (
+                  <Languages
+                    key="Languages"
+                    items={languages}
+                    value={languagesValue}
+                    getItemHref={getLanguagesItemHref}
+                    getItemLabel={getLanguagesItemLabel}
+                    getItemOnClick={getLanguagesItemOnClick}
+                    getItemTarget={getLanguagesItemTarget}
+                    onChange={onLanguageChange}
+                    style={{ zIndex: elementZIndex }}
+                  />
+                ) : (
+                  undefined
+                ),
+                additionalButtons?.length ? (
+                  <ButtonMenu
+                    items={additionalButtons}
+                    getItemHref={getAdditionalButtonsItemHref}
+                    getItemIcon={getAdditionalButtonsItemIcon}
+                    getItemLabel={getAdditionalButtonsItemLabel}
+                    getItemOnClick={getAdditionalButtonsItemOnClick}
+                    getItemTarget={getAdditionalButtonsItemTarget}
+                    onItemClick={onAdditionalButtonsItemClick}
+                    view="clear"
+                    size="s"
+                  />
+                ) : (
+                  undefined
+                ),
+              ]}
+            </div>
+          ),
+        }}
         rowCenter={{
           left: (
             <div className={cnHeader('RowCenterLeft', { breakpoint: 'm' })}>
               <HeaderLogo className={cnHeader('Logo')} logo={logo} href={logoHref} />
+              {(searchOnSubmit || searchOnChange) && (
+                <HeaderSearch
+                  value={searchValue}
+                  onChange={searchOnChange}
+                  onSubmit={searchOnSubmit}
+                  placeholder={searchPlaceholder}
+                />
+              )}
               {menu && (
                 <Menu
                   className={cnHeader('Menu')}
@@ -140,7 +275,7 @@ const HeaderRender = (props: HeaderProps, ref: React.Ref<HTMLDivElement>) => {
                   getItemHref={getTileMenuItemHref}
                   getItemImage={getTileMenuItemImage}
                   getItemOnClick={getTileMenuItemOnClick}
-                  getItemTitle={getTileMenuItemTitle}
+                  getItemLabel={getTileMenuItemLabel}
                   listClassName={cnHeader('TileMenuList', { breakpoint: 'm' })}
                   title={tileMenuTitle}
                   style={{ zIndex: elementZIndex }}
@@ -211,18 +346,42 @@ const HeaderRender = (props: HeaderProps, ref: React.Ref<HTMLDivElement>) => {
               {menu && (
                 <MobileMenu
                   className={cnHeader('Menu')}
-                  items={menu}
+                  items={getMobileMenu(props, { languagesValue, setLanguagesValue })}
                   getItemActive={getMenuItemActive}
                   getItemHref={getMenuItemHref}
                   getItemLabel={getMenuItemLabel}
                   getItemOnClick={getMenuItemOnClick}
                   getItemSubMenu={getMenuItemSubMenu}
                   getItemTarget={getMenuItemTarget}
-                  onItemClick={onMenuItemClick}
                   style={{ zIndex: elementZIndex }}
+                  footer={
+                    socialMedia?.length && (
+                      <ButtonMenu
+                        className={cnHeader('MobileMenuSocialMedia')}
+                        items={socialMedia}
+                        getItemHref={getSocialMediaItemHref}
+                        getItemIcon={getSocialMediaItemIcon}
+                        getItemLabel={getSocialMediaItemLabel}
+                        getItemOnClick={getSocialMediaItemOnClick}
+                        getItemTarget={getSocialMediaItemTarget}
+                        onItemClick={onSocialMediaItemClick}
+                        view="clear"
+                        onlyIcon
+                        size="s"
+                      />
+                    )
+                  }
                 />
               )}
               <HeaderLogo className={cnHeader('Logo')} logo={logo} href={logoHref} />
+              {(searchOnSubmit || searchOnChange) && (
+                <HeaderSearch
+                  value={searchValue}
+                  onChange={searchOnChange}
+                  onSubmit={searchOnSubmit}
+                  placeholder={searchPlaceholder}
+                />
+              )}
             </div>
           ),
           right: (
@@ -244,7 +403,7 @@ const HeaderRender = (props: HeaderProps, ref: React.Ref<HTMLDivElement>) => {
                   getItemHref={getTileMenuItemHref}
                   getItemImage={getTileMenuItemImage}
                   getItemOnClick={getTileMenuItemOnClick}
-                  getItemTitle={getTileMenuItemTitle}
+                  getItemLabel={getTileMenuItemLabel}
                   listClassName={cnHeader('TileMenuList')}
                   title={tileMenuTitle}
                   isMobile
@@ -316,14 +475,13 @@ const HeaderRender = (props: HeaderProps, ref: React.Ref<HTMLDivElement>) => {
             {menu && (
               <MobileMenu
                 className={cnHeader('Menu')}
-                items={menu}
+                items={getMobileMenu(props, { languagesValue, setLanguagesValue })}
                 getItemActive={getMenuItemActive}
                 getItemHref={getMenuItemHref}
                 getItemLabel={getMenuItemLabel}
                 getItemOnClick={getMenuItemOnClick}
                 getItemSubMenu={getMenuItemSubMenu}
                 getItemTarget={getMenuItemTarget}
-                onItemClick={onMenuItemClick}
                 style={{ zIndex: elementZIndex }}
                 header={
                   <div className={cnHeader('MobileMenuHeader')}>
@@ -348,7 +506,7 @@ const HeaderRender = (props: HeaderProps, ref: React.Ref<HTMLDivElement>) => {
                             getItemHref={getTileMenuItemHref}
                             getItemImage={getTileMenuItemImage}
                             getItemOnClick={getTileMenuItemOnClick}
-                            getItemTitle={getTileMenuItemTitle}
+                            getItemLabel={getTileMenuItemLabel}
                             listClassName={cnHeader('TileMenuList')}
                             title={tileMenuTitle}
                             isMobile
@@ -385,7 +543,34 @@ const HeaderRender = (props: HeaderProps, ref: React.Ref<HTMLDivElement>) => {
                         )}
                       </div>
                     )}
+                    <div className={cnHeader('MobileSearchContainer')}>
+                      {(searchOnSubmit || searchOnChange) && (
+                        <HeaderSearch
+                          value={searchValue}
+                          onChange={searchOnChange}
+                          onSubmit={searchOnSubmit}
+                          placeholder={searchPlaceholder}
+                        />
+                      )}
+                    </div>
                   </div>
+                }
+                footer={
+                  socialMedia?.length && (
+                    <ButtonMenu
+                      className={cnHeader('MobileMenuSocialMedia')}
+                      items={socialMedia}
+                      getItemHref={getSocialMediaItemHref}
+                      getItemIcon={getSocialMediaItemIcon}
+                      getItemLabel={getSocialMediaItemLabel}
+                      getItemOnClick={getSocialMediaItemOnClick}
+                      getItemTarget={getSocialMediaItemTarget}
+                      onItemClick={onSocialMediaItemClick}
+                      view="clear"
+                      onlyIcon
+                      size="s"
+                    />
+                  )
                 }
               />
             )}