+++ src/MobileMenu/types.ts
@@ -7,9 +7,11 @@ import {
   VerticalMenuPropOnItemClick,
   VerticalMenuPropGetItemTarget,
   VerticalMenuPropGetItemSubMenu,
+  VerticalMenuPropGetItemGroup,
 } from '@/VerticalMenu'
 
 import { PropsWithHTMLAttributesAndRef } from '@/__private__/utils/types/PropsWithHTMLAttributes'
+import React from 'react'
 
 export type MobileMenuProps<ITEM = DefaultItem> = PropsWithHTMLAttributesAndRef<
   {
@@ -20,10 +22,12 @@ export type MobileMenuProps<ITEM = DefaultItem> = PropsWithHTMLAttributesAndRef<
     getItemActive?: VerticalMenuPropGetItemActive<ITEM>
     getItemOnClick?: VerticalMenuPropGetItemOnClick<ITEM>
     getItemSubMenu?: VerticalMenuPropGetItemSubMenu<ITEM>
+    getItemGroup?: VerticalMenuPropGetItemGroup<ITEM>
     onItemClick?: VerticalMenuPropOnItemClick<ITEM>
     header?: React.ReactNode
     sidebarClassName?: string
     children?: never
+    footer?: React.ReactNode
   },
   HTMLButtonElement
 > &
@@ -34,3 +38,5 @@ export type MobileMenuProps<ITEM = DefaultItem> = PropsWithHTMLAttributesAndRef<
 export type MobileMenuComponent = <ITEM = DefaultItem>(
   props: MobileMenuProps<ITEM>
 ) => React.ReactElement | null
+
+export { DefaultItem }