+++ src/TileMenu/__stories__/examples/TileMenuExampleView.tsx
@@ -6,42 +6,42 @@ import { TileMenu, DefaultItem } from '@/TileMenu'
 
 const items: DefaultItem[] = [
   {
-    title: 'Портал',
+    label: 'Портал',
     description: 'Сводная информация обо мне и подразделении, новости компании',
     image: 'https://avatars.githubusercontent.com/u/13190808?v=4',
   },
   {
-    title: 'Портал',
+    label: 'Портал',
     description: 'Сводная информация обо мне и подразделении, новости компании',
     image: 'https://avatars.githubusercontent.com/u/13190808?v=4',
   },
   {
-    title: 'Портал',
+    label: 'Портал',
     description: 'Сводная информация обо мне и подразделении, новости компании',
     image: 'https://avatars.githubusercontent.com/u/13190808?v=4',
   },
   {
-    title: 'Портал',
+    label: 'Портал',
     description: 'Сводная информация обо мне и подразделении, новости компании',
     image: 'https://avatars.githubusercontent.com/u/13190808?v=4',
   },
   {
-    title: 'Портал',
+    label: 'Портал',
     description: 'Сводная информация обо мне и подразделении, новости компании',
     image: 'https://avatars.githubusercontent.com/u/13190808?v=4',
   },
   {
-    title: 'Портал',
+    label: 'Портал',
     description: 'Сводная информация обо мне и подразделении, новости компании',
     image: 'https://avatars.githubusercontent.com/u/13190808?v=4',
   },
   {
-    title: 'Портал',
+    label: 'Портал',
     description: 'Сводная информация обо мне и подразделении, новости компании',
     image: 'https://avatars.githubusercontent.com/u/13190808?v=4',
   },
   {
-    title: 'Портал',
+    label: 'Портал',
     description: 'Сводная информация обо мне и подразделении, новости компании',
     image: 'https://avatars.githubusercontent.com/u/13190808?v=4',
   },