+++ src/VerticalMenu/VerticalMenu.tsx
@@ -1,20 +1,23 @@
-import React, { useState, useCallback } from 'react'
+import React, { useState, useCallback, forwardRef, useEffect } from 'react'
 
 import { cn } from '@/__private__/utils/bem'
 
-import { withDefaultGetters, getItemClick } from './helpers'
-import { VerticalMenuComponent, Level } from './types'
+import { withDefaultGetters } from './helpers'
+import { getItemClick } from '@/__private__/helpers/getItemClick'
+import { VerticalMenuComponent, Level, VerticalMenuProps, DefaultItem } from './types'
 import './VerticalMenu.css'
 import { VerticalMenuLevel } from './VerticalMenuLevel/VerticalMenuLevel'
 import { TransitionGroup, CSSTransition } from 'react-transition-group'
 import { cnForCssTransition } from '@consta/uikit/__internal__/src/utils/cnForCssTransition'
 import { useDebounce } from '@consta/uikit/useDebounce'
+import { useMutableRef } from '@consta/uikit/useMutableRef'
+import { useRefs } from '@consta/uikit/useRefs'
 
 export const cnVerticalMenu = cn('VerticalMenu')
 
 const transitionCn = cnForCssTransition(cnVerticalMenu, 'Level')
 
-export const VerticalMenu: VerticalMenuComponent = props => {
+const VerticalMenuRender = (props: VerticalMenuProps, ref: React.Ref<HTMLDivElement>) => {
   const {
     items,
     className,
@@ -26,12 +29,15 @@ export const VerticalMenu: VerticalMenuComponent = props => {
     getItemSubMenu,
     header,
     onItemClick,
+    getItemGroup,
+    footer,
     ...otherProps
   } = withDefaultGetters(props)
 
-  const [levels, setLevels] = useState<Array<Level<typeof items[number]>>>([{ items }])
-
+  const [levels, setLevels] = useState<Array<Level<DefaultItem>>>([{ items, id: '0' }])
   const [animationBack, setAnimationBack] = useState<boolean>(false)
+
+  const levelRefs = useRefs<HTMLDivElement>(levels.length)
 
   const disableAnimationBack = useDebounce(() => setAnimationBack(false), 250)
 
@@ -55,15 +61,54 @@ export const VerticalMenu: VerticalMenuComponent = props => {
     disableAnimationBack()
   }, [])
 
+  const getItemSubMenuRef = useMutableRef(getItemSubMenu)
+  const getItemLabelRef = useMutableRef(getItemLabel)
+
+  useEffect(() => {
+    // обновляем оровни при смене items
+    const keys = levels[levels.length - 1].id.split('-').slice(1)
+
+    let id = '0'
+
+    const newLevels: Array<Level<DefaultItem>> = [{ items, id }]
+
+    keys.forEach(index => {
+      const item = newLevels[newLevels.length - 1].items[Number(index)]
+
+      if (!item) {
+        return
+      }
+
+      const levelItems = getItemSubMenuRef.current(item)
+      const levelLabel = getItemLabelRef.current(item)
+      id = `${id}-${index}`
+
+      if (!levelItems) {
+        return
+      }
+
+      newLevels.push({ items: levelItems, label: levelLabel, id })
+    })
+
+    setLevels(newLevels)
+  }, [items])
+
   return (
-    <div {...otherProps} className={cnVerticalMenu({ animationBack }, [className])}>
+    <div {...otherProps} className={cnVerticalMenu({ animationBack }, [className])} ref={ref}>
       <TransitionGroup enter exit className={cnVerticalMenu('Levels')}>
         {levels.map((level, index) => {
           const levelNum = index + 1
           if (levelNum >= levels.length) {
             return (
-              <CSSTransition classNames={transitionCn} key={levelNum} timeout={250}>
+              <CSSTransition
+                classNames={transitionCn}
+                key={levelNum}
+                timeout={250}
+                nodeRef={levelRefs[index]}
+              >
                 <VerticalMenuLevel
+                  ref={levelRefs[index]}
+                  id={level.id}
                   items={level.items}
                   label={level.label}
                   addLevel={addLevel}
@@ -73,12 +118,14 @@ export const VerticalMenu: VerticalMenuComponent = props => {
                   })}
                   key={levelNum}
                   header={index === 0 && header}
+                  footer={index === 0 && footer}
                   getItemActive={getItemActive}
                   getItemHref={getItemHref}
                   getItemLabel={getItemLabel}
                   getItemOnClick={menuItem => getItemClick(menuItem, getItemOnClick, onItemClick)}
                   getItemTarget={getItemTarget}
                   getItemSubMenu={getItemSubMenu}
+                  getItemGroup={getItemGroup}
                 />
               </CSSTransition>
             )
@@ -88,5 +135,7 @@ export const VerticalMenu: VerticalMenuComponent = props => {
     </div>
   )
 }
+
+export const VerticalMenu = forwardRef(VerticalMenuRender) as VerticalMenuComponent
 
 export * from './types'